+++ tests/testthat/test-screening.R
@@ -4,7 +4,7 @@ susps <- as.data.table(patRoonData::targets)
 
 susps[, InChI := babelConvert(SMILES, "smi", "inchi")]
 susps[, neutralMass := getNeutralMassFromSMILES(SMILES)]
-susps[, formula := sapply(getMoleculesFromSMILES(SMILES), function(m) rcdk::get.mol2formula(m)@string)]
+susps[, formula := convertToFormulaBabel(SMILES, "smi")]
 susps[, adduct := "[M+H]+"]
 susps[name %in% c("TBA", "TPA"), adduct := "[M]+"]
 
@@ -46,12 +46,14 @@ test_that("suspect screening is OK", {
     expect_equal(scrSMI, screenSuspects(fGroups, suspsMissing[, c("name", "rt", "adduct", "formula", "SMILES")]))
     expect_equal(scrSMI, screenSuspects(fGroups, suspsMissing[, c("name", "rt", "adduct", "SMILES", "InChI")]))
     
-    expect_warning(screenSuspects(fGroups, suspsMissingRow, skipInvalid = TRUE))
-    expect_error(screenSuspects(fGroups, suspsMissingRow, skipInvalid = FALSE))
-    
     # adduct argument
     expect_equal(screenSuspects(fGroups, susps[name %in% c("TBA", "TPA")]),
                  screenSuspects(fGroups, susps[name %in% c("TBA", "TPA"), -"adduct"], adduct = "[M]+"))
+
+    # certain Linux/JVM combinations (e.g. current CI Docker image) give Java StackOverflows...
+    testthat::skip_on_os("linux")
+    expect_warning(screenSuspects(fGroups, suspsMissingRow, skipInvalid = TRUE))
+    expect_error(screenSuspects(fGroups, suspsMissingRow, skipInvalid = FALSE))
 })
 
 TQFile <- file.path(getTestDataPath(), "GlobalResults-TASQ.csv")