+++ NEWS.md
@@ -11,6 +11,8 @@
     * More extra commandline options can be given to `SIRIUS`
 * Fixed: `groupNames()`, `analyses()` and similar methods sometimes returned `NULL` instead of an empty `character` vector for empty objects.
 * `plotHeatMap()` with `interactive=TRUE`: switch from now removed `d3heatmap` package to `heatmaply`
+* Fixed: `reportHTML()` didn't split PubChem URLs when multiple identifiers were reported.
+
 
 ## April 2020